requirements.txt
scipy
numpy
matplotlib
Pillow


tensorflow